✅ Task 5完了: PerformanceDiagnostics.js分割 - Main Controller Pattern適用
Phase F.3 Issue #95対応の一環として、PerformanceDiagnostics.js (2,644語)をMain Controller Patternで分割:

**分割構成**:
• DiagnosticDataCollector.js (962語) - データ収集・監視機能
• DiagnosticAnalyzer.js (1,681語) - 診断分析・問題検出機能
• DiagnosticReporter.js (1,193語) - 報告・結果表示機能
• PerformanceDiagnostics.js (782語) - メインコントローラー

**達成事項**:
• 70%サイズ削減 (2,644語→782語)
• 完全な後方互換性維持 (既存API委譲による保持)
• 依存注入パターンでコンポーネント統制
• 全3コンポーネントが2,500語以下のMCPツール互換性

**MCPツール対応**:
find_symbol等の25,000トークン制限問題解決、開発効率向上
