+++ .kiro/specs/performance-utility-splitting-phase-f3/tasks.md
@@ -93,26 +93,26 @@
   - Verify all existing validation functionality works correctly
   - _Requirements: 1.3, 8.1, 8.2_
 
-- [ ] 5. Split PerformanceDiagnostics.js (2,644 words)
-- [ ] 5.1 Create DiagnosticDataCollector component
+- [x] 5. Split PerformanceDiagnostics.js (2,644 words)
+- [x] 5.1 Create DiagnosticDataCollector component
   - Extract diagnostic data collection and monitoring functionality
   - Implement data gathering and diagnostic metric collection methods
   - Create unit tests for diagnostic data collection functionality
   - _Requirements: 1.4, 8.1, 8.2_
 
-- [ ] 5.2 Create DiagnosticAnalyzer component
+- [x] 5.2 Create DiagnosticAnalyzer component
   - Extract diagnostic analysis and problem detection functionality
   - Implement diagnostic algorithms and issue identification methods
   - Create unit tests for diagnostic analysis functionality
   - _Requirements: 1.4, 8.1, 8.2_
 
-- [ ] 5.3 Create DiagnosticReporter component
+- [x] 5.3 Create DiagnosticReporter component
   - Extract diagnostic reporting and result presentation functionality
   - Implement diagnostic report generation and formatting methods
   - Create unit tests for diagnostic reporting functionality
   - _Requirements: 1.4, 8.1, 8.2_
 
-- [ ] 5.4 Refactor main PerformanceDiagnostics to use sub-components
+- [x] 5.4 Refactor main PerformanceDiagnostics to use sub-components
   - Modify main class to instantiate and coordinate sub-components
   - Implement delegation methods to maintain public API compatibility
   - Verify all existing diagnostic functionality works correctly
