+++ src/utils/PerformanceDiagnostics.js
@@ -1,38 +1,38 @@
 /**
- * Performance Diagnostics System
- * 
+ * PerformanceDiagnostics - Main Controller for performance diagnostics system
+ * Refactored to use the Main Controller Pattern with sub-components
  * パフォーマンスボトルネック特定と自動診断システム
  * Requirements: 7.2, 7.4
  */
 
+import { DiagnosticDataCollector } from './performance-diagnostics/DiagnosticDataCollector.js';
+import { DiagnosticAnalyzer } from './performance-diagnostics/DiagnosticAnalyzer.js';
+import { DiagnosticReporter } from './performance-diagnostics/DiagnosticReporter.js';
+
 export class PerformanceDiagnostics {
     constructor() {
-        this.diagnosticEngine = new DiagnosticEngine();
-        this.bottleneckIdentifier = new BottleneckIdentifier();
-        this.issueAnalyzer = new IssueAnalyzer();
-        this.recommendationEngine = new RecommendationEngine();
-        this.reportGenerator = new DiagnosticReportGenerator();
-        this.anomalyDetector = new AnomalyDetector();
-        this.rootCauseAnalyzer = new RootCauseAnalyzer();
+        // Initialize sub-components with dependency injection
+        this.dataCollector = new DiagnosticDataCollector(this);
+        this.analyzer = new DiagnosticAnalyzer(this);
+        this.reporter = new DiagnosticReporter(this);
+        
         this.initialized = false;
         
         this.initializeDiagnostics();
+        
+        console.log('[PerformanceDiagnostics] Main controller initialized successfully');
     }
 
     async initializeDiagnostics() {
         try {
-            await this.diagnosticEngine.initialize();
-            await this.bottleneckIdentifier.initialize();
-            await this.issueAnalyzer.initialize();
-            await this.recommendationEngine.initialize();
-            await this.reportGenerator.initialize();
-            await this.anomalyDetector.initialize();
-            await this.rootCauseAnalyzer.initialize();
+            // Initialize sub-components
+            await this.analyzer.initialize();
+            await this.reporter.initialize();
             
             this.initialized = true;
-            console.log('PerformanceDiagnostics initialized successfully');
+            console.log('[PerformanceDiagnostics] Main controller initialized successfully');
         } catch (error) {
-            console.error('Failed to initialize PerformanceDiagnostics:', error);
+            console.error('[PerformanceDiagnostics] Failed to initialize:', error);
             throw error;
         }
     }
@@ -58,11 +58,16 @@ export class PerformanceDiagnostics {
         };
 
         try {
-            // データ収集期間
-            const collectedData = await this.collectDiagnosticData(diagnosticSession.options);
+            // データ収集期間 - delegated to data collector
+            const collectedData = await this.dataCollector.collectDiagnosticData(diagnosticSession.options);
             
-            // 各種分析の実行
-            const analysisResults = await this.runAnalyses(collectedData, diagnosticSession.options);
+            // 各種分析の実行 - delegated to analyzer
+            const analysisResults = await this.analyzer.runAnalyses(collectedData, diagnosticSession.options);
+            
+            // 推奨事項生成 - delegated to reporter
+            if (diagnosticSession.options.generateRecommendations) {
+                analysisResults.recommendations = await this.reporter.generateRecommendations(analysisResults);
+            }
             
             // 結果の統合
             diagnosticSession.results = {
@@ -71,14 +76,14 @@ export class PerformanceDiagnostics {
                 anomalies: analysisResults.anomalies,
                 rootCauses: analysisResults.rootCauses,
                 recommendations: analysisResults.recommendations,
-                overallAssessment: await this.generateOverallAssessment(analysisResults)
+                overallAssessment: await this.analyzer.generateOverallAssessment(analysisResults)
             };
 
             diagnosticSession.endTime = performance.now();
             diagnosticSession.duration = diagnosticSession.endTime - diagnosticSession.startTime;
 
-            // 診断レポート生成
-            const report = await this.reportGenerator.generate(diagnosticSession);
+            // 診断レポート生成 - delegated to reporter
+            const report = await this.reporter.generateReport(diagnosticSession);
             
             return {
                 session: diagnosticSession,
@@ -86,220 +91,48 @@ export class PerformanceDiagnostics {
             };
 
         } catch (error) {
-            console.error('Comprehensive diagnosis failed:', error);
+            console.error('[PerformanceDiagnostics] Comprehensive diagnosis failed:', error);
             throw error;
         }
     }
 
+    // ===== DELEGATED METHODS - Maintain backward compatibility =====
+    
+    /**
+     * Collect diagnostic data - delegated to data collector
+     * @param {Object} options - Collection options
+     * @returns {Object} Collected data
+     */
     async collectDiagnosticData(options) {
-        console.log('Collecting diagnostic data...');
-        
-        const dataCollector = new DiagnosticDataCollector();
-        await dataCollector.start(options);
-        
-        // データ収集期間の待機
-        await new Promise(resolve => setTimeout(resolve, options.duration));
-        
-        const collectedData = await dataCollector.stop();
-        
-        return {
-            rawData: collectedData,
-            summary: this.summarizeCollectedData(collectedData)
-        };
+        return await this.dataCollector.collectDiagnosticData(options);
     }
 
-    summarizeCollectedData(data) {
-        return {
-            duration: data.collectionDuration,
-            sampleCount: data.samples.length,
-            metricsCollected: Object.keys(data.metrics).length,
-            timeRange: {
-                start: data.startTime,
-                end: data.endTime
-            },
-            dataQuality: this.assessDataQuality(data)
-        };
-    }
-
-    assessDataQuality(data) {
-        const completeness = data.samples.filter(s => s.complete).length / data.samples.length;
-        const consistency = this.calculateDataConsistency(data.samples);
-        
-        return {
-            completeness: completeness,
-            consistency: consistency,
-            overall: (completeness + consistency) / 2,
-            issues: this.identifyDataQualityIssues(data)
-        };
-    }
-
-    calculateDataConsistency(samples) {
-        // データの一貫性を計算
-        if (samples.length < 2) return 1.0;
-        
-        let consistencyScore = 0;
-        const expectedInterval = samples[1].timestamp - samples[0].timestamp;
-        
-        for (let i = 1; i < samples.length; i++) {
-            const actualInterval = samples[i].timestamp - samples[i - 1].timestamp;
-            const deviation = Math.abs(actualInterval - expectedInterval) / expectedInterval;
-            consistencyScore += Math.max(0, 1 - deviation);
-        }
-        
-        return consistencyScore / (samples.length - 1);
-    }
-
-    identifyDataQualityIssues(data) {
-        const issues = [];
-        
-        // 欠損データの検出
-        const incompleteSamples = data.samples.filter(s => !s.complete);
-        if (incompleteSamples.length > data.samples.length * 0.1) {
-            issues.push({
-                type: 'missing_data',
-                severity: 'medium',
-                description: `${incompleteSamples.length} incomplete samples detected`,
-                impact: 'May affect analysis accuracy'
-            });
-        }
-        
-        // 異常な時間間隔の検出
-        const timeGaps = this.detectTimeGaps(data.samples);
-        if (timeGaps.length > 0) {
-            issues.push({
-                type: 'time_gaps',
-                severity: 'low',
-                description: `${timeGaps.length} time gaps detected`,
-                impact: 'May indicate system overload during collection'
-            });
-        }
-        
-        return issues;
-    }
-
-    detectTimeGaps(samples) {
-        const gaps = [];
-        const expectedInterval = 100; // 100ms expected interval
-        
-        for (let i = 1; i < samples.length; i++) {
-            const interval = samples[i].timestamp - samples[i - 1].timestamp;
-            if (interval > expectedInterval * 2) {
-                gaps.push({
-                    start: samples[i - 1].timestamp,
-                    end: samples[i].timestamp,
-                    duration: interval
-                });
-            }
-        }
-        
-        return gaps;
-    }
-
+    /**
+     * Run analyses - delegated to analyzer
+     * @param {Object} collectedData - Collected data
+     * @param {Object} options - Analysis options
+     * @returns {Object} Analysis results
+     */
     async runAnalyses(collectedData, options) {
-        const results = {};
-        
-        // ボトルネック分析
-        if (options.includeBottleneckAnalysis) {
-            console.log('Running bottleneck analysis...');
-            results.bottlenecks = await this.bottleneckIdentifier.analyze(collectedData);
-        }
-        
-        // 異常検出
-        if (options.includeAnomalyDetection) {
-            console.log('Running anomaly detection...');
-            results.anomalies = await this.anomalyDetector.detect(collectedData);
-        }
-        
-        // 根本原因分析
-        if (options.includeRootCauseAnalysis) {
-            console.log('Running root cause analysis...');
-            results.rootCauses = await this.rootCauseAnalyzer.analyze(collectedData, results);
-        }
-        
-        // 推奨事項生成
-        if (options.generateRecommendations) {
-            console.log('Generating recommendations...');
-            results.recommendations = await this.recommendationEngine.generate(results);
-        }
-        
-        return results;
+        return await this.analyzer.runAnalyses(collectedData, options);
     }
 
+    /**
+     * Generate overall assessment - delegated to analyzer
+     * @param {Object} analysisResults - Analysis results
+     * @returns {Object} Overall assessment
+     */
     async generateOverallAssessment(analysisResults) {
-        const assessment = {
-            healthScore: 100,
-            performanceLevel: 'excellent',
-            criticalIssues: [],
-            majorConcerns: [],
-            improvements: [],
-            summary: ''
-        };
-
-        // ボトルネック評価
-        if (analysisResults.bottlenecks) {
-            const criticalBottlenecks = analysisResults.bottlenecks.filter(b => b.severity === 'critical');
-            const majorBottlenecks = analysisResults.bottlenecks.filter(b => b.severity === 'high');
-            
-            assessment.healthScore -= criticalBottlenecks.length * 20;
-            assessment.healthScore -= majorBottlenecks.length * 10;
-            
-            assessment.criticalIssues.push(...criticalBottlenecks.map(b => ({
-                type: 'bottleneck',
-                description: b.description,
-                component: b.component,
-                impact: b.impact
-            })));
-        }
-
-        // 異常評価
-        if (analysisResults.anomalies) {
-            const criticalAnomalies = analysisResults.anomalies.filter(a => a.severity === 'critical');
-            const majorAnomalies = analysisResults.anomalies.filter(a => a.severity === 'high');
-            
-            assessment.healthScore -= criticalAnomalies.length * 15;
-            assessment.healthScore -= majorAnomalies.length * 8;
-            
-            assessment.criticalIssues.push(...criticalAnomalies.map(a => ({
-                type: 'anomaly',
-                description: a.description,
-                metric: a.metric,
-                deviation: a.deviation
-            })));
-        }
-
-        // パフォーマンスレベル決定
-        assessment.healthScore = Math.max(0, assessment.healthScore);
-        if (assessment.healthScore >= 90) assessment.performanceLevel = 'excellent';
-        else if (assessment.healthScore >= 75) assessment.performanceLevel = 'good';
-        else if (assessment.healthScore >= 60) assessment.performanceLevel = 'fair';
-        else if (assessment.healthScore >= 40) assessment.performanceLevel = 'poor';
-        else assessment.performanceLevel = 'critical';
-
-        // サマリー生成
-        assessment.summary = this.generateAssessmentSummary(assessment);
-
-        return assessment;
+        return await this.analyzer.generateOverallAssessment(analysisResults);
     }
 
-    generateAssessmentSummary(assessment) {
-        const level = assessment.performanceLevel;
-        const score = assessment.healthScore;
-        const issues = assessment.criticalIssues.length;
-        
-        if (level === 'excellent') {
-            return `システムは優秀な状態です（スコア: ${score}）。大きな問題は検出されていません。`;
-        } else if (level === 'good') {
-            return `システムは良好な状態です（スコア: ${score}）。軽微な改善の余地があります。`;
-        } else if (level === 'fair') {
-            return `システムは普通の状態です（スコア: ${score}）。${issues}件の問題が検出されました。`;
-        } else if (level === 'poor') {
-            return `システムに性能問題があります（スコア: ${score}）。${issues}件の重要な問題への対処が必要です。`;
-        } else {
-            return `システムは深刻な状態です（スコア: ${score}）。${issues}件の致命的な問題への緊急対処が必要です。`;
-        }
-    }
-
-    // 単発診断機能
+    // ===== PUBLIC API METHODS - Enhanced with sub-component functionality =====
+    
+    /**
+     * Quick diagnosis with basic analysis
+     * @param {string} targetMetric - Optional target metric to focus on
+     * @returns {Object} Quick diagnosis results
+     */
     async quickDiagnosis(targetMetric = null) {
         const options = {
             duration: 5000, // 5秒
@@ -314,795 +147,138 @@ export class PerformanceDiagnostics {
         return await this.runComprehensiveDiagnosis(options);
     }
 
+    /**
+     * Identify bottlenecks only
+     * @param {number} duration - Collection duration in milliseconds
+     * @returns {Array} Identified bottlenecks
+     */
     async identifyBottlenecks(duration = 10000) {
-        const options = {
-            duration,
-            includeAnomalyDetection: false,
-            includeRootCauseAnalysis: false,
-            detailLevel: 'standard'
-        };
-
-        const result = await this.runComprehensiveDiagnosis(options);
-        return result.session.results.bottlenecks;
+        const collectedData = await this.dataCollector.collectDiagnosticData({ duration });
+        return await this.analyzer.identifyBottlenecks(collectedData);
     }
 
+    /**
+     * Detect anomalies only
+     * @param {number} duration - Collection duration in milliseconds
+     * @returns {Array} Detected anomalies
+     */
     async detectAnomalies(duration = 15000) {
-        const options = {
-            duration,
-            includeBottleneckAnalysis: false,
-            includeRootCauseAnalysis: false,
-            detailLevel: 'standard'
-        };
-
-        const result = await this.runComprehensiveDiagnosis(options);
-        return result.session.results.anomalies;
+        const collectedData = await this.dataCollector.collectDiagnosticData({ duration });
+        return await this.analyzer.detectAnomalies(collectedData);
     }
 
-    // 公開API
+    /**
+     * Main diagnosis API - alias for runComprehensiveDiagnosis
+     * @param {Object} options - Diagnosis options
+     * @returns {Object} Diagnosis results
+     */
     async diagnose(options = {}) {
         return await this.runComprehensiveDiagnosis(options);
     }
 
+    /**
+     * Get current system health assessment
+     * @returns {Object} System health assessment
+     */
     async getSystemHealth() {
         const quickResult = await this.quickDiagnosis();
         return quickResult.session.results.overallAssessment;
     }
 
+    /**
+     * Get performance recommendations
+     * @returns {Array} Performance recommendations
+     */
     async getRecommendations() {
         const result = await this.runComprehensiveDiagnosis({ duration: 10000 });
         return result.session.results.recommendations;
     }
-}
 
-// 診断エンジン
-class DiagnosticEngine {
-    constructor() {
-        this.processors = new Map();
-        this.ruleEngine = new DiagnosticRuleEngine();
-    }
-
-    async initialize() {
-        await this.setupProcessors();
-        await this.ruleEngine.initialize();
-    }
-
-    async setupProcessors() {
-        this.processors.set('frame_analysis', new FrameAnalysisProcessor());
-        this.processors.set('memory_analysis', new MemoryAnalysisProcessor());
-        this.processors.set('render_analysis', new RenderAnalysisProcessor());
-        this.processors.set('network_analysis', new NetworkAnalysisProcessor());
-        this.processors.set('interaction_analysis', new InteractionAnalysisProcessor());
-
-        for (const processor of this.processors.values()) {
-            await processor.initialize();
-        }
-    }
-
-    async processData(data, analysisType) {
-        const processor = this.processors.get(analysisType);
-        if (!processor) {
-            throw new Error(`Unknown analysis type: ${analysisType}`);
-        }
-
-        return await processor.process(data);
-    }
-
-    async applyRules(data, context) {
-        return await this.ruleEngine.apply(data, context);
-    }
-}
-
-// ボトルネック特定器
-class BottleneckIdentifier {
-    constructor() {
-        this.detectors = new Map();
-        this.thresholds = new Map();
-    }
-
-    async initialize() {
-        this.setupDetectors();
-        this.setupThresholds();
-    }
-
-    setupDetectors() {
-        this.detectors.set('frame_rate', new FrameRateBottleneckDetector());
-        this.detectors.set('memory', new MemoryBottleneckDetector());
-        this.detectors.set('rendering', new RenderingBottleneckDetector());
-        this.detectors.set('network', new NetworkBottleneckDetector());
-        this.detectors.set('computation', new ComputationBottleneckDetector());
-    }
-
-    setupThresholds() {
-        this.thresholds.set('frame_rate', {
-            critical: 20, // 20fps以下は致命的
-            high: 40,     // 40fps以下は高優先度
-            medium: 55    // 55fps以下は中優先度
-        });
-
-        this.thresholds.set('memory_growth', {
-            critical: 5 * 1024 * 1024, // 5MB/s以上の成長は致命的
-            high: 2 * 1024 * 1024,     // 2MB/s以上は高優先度
-            medium: 1 * 1024 * 1024    // 1MB/s以上は中優先度
-        });
-
-        this.thresholds.set('render_time', {
-            critical: 50, // 50ms以上は致命的
-            high: 33.33,  // 30fps相当以上は高優先度
-            medium: 20    // 20ms以上は中優先度
-        });
-    }
-
-    async analyze(collectedData) {
-        const bottlenecks = [];
-
-        for (const [detectorName, detector] of this.detectors) {
-            try {
-                const detected = await detector.detect(collectedData, this.thresholds);
-                bottlenecks.push(...detected);
-            } catch (error) {
-                console.error(`Bottleneck detection failed for ${detectorName}:`, error);
-            }
-        }
-
-        // 重要度でソート
-        return bottlenecks.sort((a, b) => {
-            const severityOrder = { critical: 4, high: 3, medium: 2, low: 1 };
-            return (severityOrder[b.severity] || 0) - (severityOrder[a.severity] || 0);
-        });
-    }
-}
-
-// 問題分析器
-class IssueAnalyzer {
-    constructor() {
-        this.analyzers = new Map();
-    }
-
-    async initialize() {
-        this.analyzers.set('performance_degradation', new PerformanceDegradationAnalyzer());
-        this.analyzers.set('resource_contention', new ResourceContentionAnalyzer());
-        this.analyzers.set('inefficient_algorithms', new InefficientAlgorithmAnalyzer());
-        this.analyzers.set('memory_leaks', new MemoryLeakAnalyzer());
-        this.analyzers.set('excessive_garbage_collection', new ExcessiveGCAnalyzer());
-    }
-
-    async analyzeIssues(data, context) {
-        const issues = [];
-
-        for (const [analyzerName, analyzer] of this.analyzers) {
-            try {
-                const detected = await analyzer.analyze(data, context);
-                issues.push(...detected);
-            } catch (error) {
-                console.error(`Issue analysis failed for ${analyzerName}:`, error);
-            }
-        }
-
-        return issues;
-    }
-}
-
-// 推奨事項エンジン
-class RecommendationEngine {
-    constructor() {
-        this.generators = new Map();
-        this.knowledgeBase = new PerformanceKnowledgeBase();
-    }
-
-    async initialize() {
-        await this.knowledgeBase.initialize();
-        this.setupGenerators();
-    }
-
-    setupGenerators() {
-        this.generators.set('optimization', new OptimizationRecommendationGenerator());
-        this.generators.set('configuration', new ConfigurationRecommendationGenerator());
-        this.generators.set('architecture', new ArchitectureRecommendationGenerator());
-        this.generators.set('monitoring', new MonitoringRecommendationGenerator());
-    }
-
-    async generate(analysisResults) {
-        const recommendations = [];
-
-        // ボトルネックベースの推奨事項
-        if (analysisResults.bottlenecks) {
-            for (const bottleneck of analysisResults.bottlenecks) {
-                const rec = await this.generateBottleneckRecommendations(bottleneck);
-                recommendations.push(...rec);
-            }
-        }
-
-        // 異常ベースの推奨事項
-        if (analysisResults.anomalies) {
-            for (const anomaly of analysisResults.anomalies) {
-                const rec = await this.generateAnomalyRecommendations(anomaly);
-                recommendations.push(...rec);
-            }
-        }
-
-        // 一般的な推奨事項
-        const generalRec = await this.generateGeneralRecommendations(analysisResults);
-        recommendations.push(...generalRec);
-
-        return this.prioritizeAndDeduplicateRecommendations(recommendations);
-    }
-
-    async generateBottleneckRecommendations(bottleneck) {
-        const recommendations = [];
-
-        switch (bottleneck.type) {
-            case 'frame_rate':
-                recommendations.push({
-                    type: 'optimization',
-                    priority: 'high',
-                    category: 'rendering',
-                    title: 'フレームレート最適化',
-                    description: 'レンダリングパイプラインの最適化によりフレームレートを改善',
-                    actions: [
-                        'AdaptiveQualityController の設定確認',
-                        'レンダリング負荷の軽減',
-                        'フレームペーシングの調整',
-                        'VSync設定の最適化'
-                    ],
-                    estimatedImpact: 'high',
-                    implementationEffort: 'medium',
-                    timeToImplement: '1-3 days'
-                });
-                break;
-
-            case 'memory':
-                recommendations.push({
-                    type: 'memory',
-                    priority: 'high',
-                    category: 'memory_management',
-                    title: 'メモリ使用量最適化',
-                    description: 'メモリ管理の改善により安定性を向上',
-                    actions: [
-                        'MemoryManager の設定見直し',
-                        'オブジェクトプールの効率化',
-                        'メモリリークの修正',
-                        'ガベージコレクション最適化'
-                    ],
-                    estimatedImpact: 'high',
-                    implementationEffort: 'medium',
-                    timeToImplement: '2-5 days'
-                });
-                break;
-
-            case 'rendering':
-                recommendations.push({
-                    type: 'rendering',
-                    priority: 'medium',
-                    category: 'rendering_optimization',
-                    title: 'レンダリング効率改善',
-                    description: 'レンダリング処理の最適化',
-                    actions: [
-                        'ダーティリージョン管理の改善',
-                        'バッチレンダリングの実装',
-                        'レイヤー構成の最適化',
-                        'シェーダー最適化'
-                    ],
-                    estimatedImpact: 'medium',
-                    implementationEffort: 'medium',
-                    timeToImplement: '3-7 days'
-                });
-                break;
-        }
-
-        return recommendations;
-    }
-
-    async generateAnomalyRecommendations(anomaly) {
-        return [{
-            type: 'investigation',
-            priority: 'medium',
-            category: 'anomaly_resolution',
-            title: `${anomaly.metric}異常の調査`,
-            description: `${anomaly.metric}で検出された異常パターンの調査と対策`,
-            actions: [
-                '異常発生パターンの詳細分析',
-                '関連システムとの相関調査',
-                '閾値設定の見直し',
-                '予防的監視の強化'
-            ],
-            estimatedImpact: 'medium',
-            implementationEffort: 'low',
-            timeToImplement: '1-2 days'
-        }];
-    }
-
-    async generateGeneralRecommendations(analysisResults) {
-        return [
-            {
-                type: 'monitoring',
-                priority: 'low',
-                category: 'continuous_improvement',
-                title: '継続的パフォーマンス監視',
-                description: '定期的なパフォーマンス監視体制の確立',
-                actions: [
-                    'PerformanceMonitoringSystem の定期実行',
-                    'アラート閾値の定期見直し',
-                    'パフォーマンストレンドの分析',
-                    'ベースライン性能の更新'
-                ],
-                estimatedImpact: 'low',
-                implementationEffort: 'low',
-                timeToImplement: '1 day'
-            }
-        ];
-    }
-
-    prioritizeAndDeduplicateRecommendations(recommendations) {
-        // 重複除去
-        const unique = recommendations.filter((rec, index, array) => 
-            array.findIndex(r => r.title === rec.title) === index
-        );
-
-        // 優先度でソート
-        const priorityOrder = { high: 3, medium: 2, low: 1 };
-        return unique.sort((a, b) => 
-            (priorityOrder[b.priority] || 0) - (priorityOrder[a.priority] || 0)
-        );
-    }
-}
-
-// 診断レポート生成器
-class DiagnosticReportGenerator {
-    constructor() {
-        this.templates = new Map();
-    }
-
-    async initialize() {
-        this.setupTemplates();
-    }
-
-    setupTemplates() {
-        this.templates.set('comprehensive', new ComprehensiveReportTemplate());
-        this.templates.set('summary', new SummaryReportTemplate());
-        this.templates.set('technical', new TechnicalReportTemplate());
-    }
-
-    async generate(diagnosticSession) {
-        const templateName = diagnosticSession.options.detailLevel || 'comprehensive';
-        const template = this.templates.get(templateName) || this.templates.get('comprehensive');
-
-        return await template.generate(diagnosticSession);
-    }
-}
-
-// データ収集器
-class DiagnosticDataCollector {
-    constructor() {
-        this.collecting = false;
-        this.samples = [];
-        this.metrics = {};
-        this.startTime = null;
-        this.endTime = null;
-    }
-
-    async start(options) {
-        this.collecting = true;
-        this.samples = [];
-        this.metrics = {};
-        this.startTime = performance.now();
-
-        // 定期的なデータ収集
-        this.collectionInterval = setInterval(() => {
-            this.collectSample();
-        }, 100); // 100ms間隔
-
-        console.log('Diagnostic data collection started');
-    }
-
-    async stop() {
-        this.collecting = false;
-        this.endTime = performance.now();
-
-        if (this.collectionInterval) {
-            clearInterval(this.collectionInterval);
-        }
-
-        console.log(`Diagnostic data collection stopped. Collected ${this.samples.length} samples`);
-
+    /**
+     * Get diagnostic capabilities
+     * @returns {Object} Available diagnostic capabilities
+     */
+    getDiagnosticCapabilities() {
         return {
-            samples: this.samples,
-            metrics: this.metrics,
-            startTime: this.startTime,
-            endTime: this.endTime,
-            collectionDuration: this.endTime - this.startTime
+            dataCollection: this.dataCollector.getCollectionStatus ? 
+                this.dataCollector.getCollectionStatus() : { available: true },
+            analysis: this.analyzer.getAnalysisCapabilities ? 
+                this.analyzer.getAnalysisCapabilities() : { available: true },
+            reporting: this.reporter.getReportingCapabilities ? 
+                this.reporter.getReportingCapabilities() : { available: true }
         };
     }
 
-    collectSample() {
-        if (!this.collecting) return;
-
-        const timestamp = performance.now();
-        const sample = {
-            timestamp,
-            complete: true,
-            frameRate: this.getCurrentFrameRate(),
-            memoryUsage: this.getCurrentMemoryUsage(),
-            renderTime: this.getCurrentRenderTime(),
-            networkLatency: this.getCurrentNetworkLatency(),
-            inputLag: this.getCurrentInputLag()
-        };
-
-        this.samples.push(sample);
-
-        // メトリクス統計の更新
-        this.updateMetricsStatistics(sample);
-    }
-
-    getCurrentFrameRate() {
-        // 現在のフレームレートを取得（簡易実装）
-        if (this.lastFrameTime) {
-            const frameTime = performance.now() - this.lastFrameTime;
-            this.lastFrameTime = performance.now();
-            return frameTime > 0 ? 1000 / frameTime : 0;
+    /**
+     * Configure diagnostic components
+     * @param {Object} config - Configuration options
+     */
+    configure(config) {
+        if (config.dataCollection) {
+            this.dataCollector.configure(config.dataCollection);
         }
-        this.lastFrameTime = performance.now();
-        return 60; // デフォルト値
-    }
-
-    getCurrentMemoryUsage() {
-        if (performance.memory) {
-            return performance.memory.usedJSHeapSize;
-        }
-        return 0;
-    }
-
-    getCurrentRenderTime() {
-        // レンダリング時間の測定（模擬実装）
-        return Math.random() * 20 + 5; // 5-25ms
-    }
-
-    getCurrentNetworkLatency() {
-        // ネットワークレイテンシの測定（模擬実装）
-        return Math.random() * 100 + 10; // 10-110ms
-    }
-
-    getCurrentInputLag() {
-        // 入力遅延の測定（模擬実装）
-        return Math.random() * 30 + 5; // 5-35ms
-    }
-
-    updateMetricsStatistics(sample) {
-        for (const [key, value] of Object.entries(sample)) {
-            if (key === 'timestamp' || key === 'complete') continue;
-
-            if (!this.metrics[key]) {
-                this.metrics[key] = {
-                    values: [],
-                    min: value,
-                    max: value,
-                    sum: 0,
-                    count: 0
-                };
-            }
-
-            const metric = this.metrics[key];
-            metric.values.push(value);
-            metric.min = Math.min(metric.min, value);
-            metric.max = Math.max(metric.max, value);
-            metric.sum += value;
-            metric.count++;
-            metric.average = metric.sum / metric.count;
-
-            // 最新100値のみ保持
-            if (metric.values.length > 100) {
-                metric.values.shift();
-            }
-        }
-    }
-}
-
-// 異常検出器
-class AnomalyDetector {
-    constructor() {
-        this.detectors = new Map();
-        this.baselines = new Map();
-    }
-
-    async initialize() {
-        this.setupDetectors();
-        this.loadBaselines();
-    }
-
-    setupDetectors() {
-        this.detectors.set('statistical', new StatisticalAnomalyDetector());
-        this.detectors.set('threshold', new ThresholdAnomalyDetector());
-        this.detectors.set('pattern', new PatternAnomalyDetector());
-        this.detectors.set('trend', new TrendAnomalyDetector());
-    }
-
-    loadBaselines() {
-        // ベースライン値の設定
-        this.baselines.set('frameRate', { mean: 60, stdDev: 5 });
-        this.baselines.set('memoryUsage', { mean: 50 * 1024 * 1024, stdDev: 10 * 1024 * 1024 });
-        this.baselines.set('renderTime', { mean: 12, stdDev: 3 });
-        this.baselines.set('networkLatency', { mean: 50, stdDev: 20 });
-        this.baselines.set('inputLag', { mean: 15, stdDev: 5 });
-    }
-
-    async detect(collectedData) {
-        const anomalies = [];
-
-        for (const [detectorName, detector] of this.detectors) {
-            try {
-                const detected = await detector.detect(collectedData, this.baselines);
-                anomalies.push(...detected);
-            } catch (error) {
-                console.error(`Anomaly detection failed for ${detectorName}:`, error);
-            }
-        }
-
-        return anomalies.sort((a, b) => {
-            const severityOrder = { critical: 4, high: 3, medium: 2, low: 1 };
-            return (severityOrder[b.severity] || 0) - (severityOrder[a.severity] || 0);
-        });
-    }
-}
-
-// 根本原因分析器
-class RootCauseAnalyzer {
-    constructor() {
-        this.correlationAnalyzer = new CorrelationAnalyzer();
-        this.causalityDetector = new CausalityDetector();
-        this.dependencyMapper = new DependencyMapper();
-    }
-
-    async initialize() {
-        await this.correlationAnalyzer.initialize();
-        await this.causalityDetector.initialize();
-        await this.dependencyMapper.initialize();
-    }
-
-    async analyze(collectedData, analysisResults) {
-        const rootCauses = [];
-
-        // 相関分析
-        const correlations = await this.correlationAnalyzer.analyze(collectedData);
         
-        // 因果関係検出
-        const causalities = await this.causalityDetector.detect(collectedData, analysisResults);
-        
-        // 依存関係マッピング
-        const dependencies = await this.dependencyMapper.map(analysisResults);
-
-        // 根本原因の特定
-        for (const bottleneck of analysisResults.bottlenecks || []) {
-            const cause = await this.identifyRootCause(bottleneck, correlations, causalities, dependencies);
-            if (cause) {
-                rootCauses.push(cause);
-            }
+        if (config.analysis) {
+            this.analyzer.configure(config.analysis);
         }
-
-        return rootCauses;
-    }
-
-    async identifyRootCause(issue, correlations, causalities, dependencies) {
-        // 簡易的な根本原因特定
-        return {
-            issue: issue.description,
-            potentialCauses: [
-                'システム負荷の増大',
-                '不適切な設定値',
-                'リソースの枯渇',
-                'アルゴリズムの非効率性'
-            ],
-            confidence: 0.7,
-            recommendedActions: [
-                '詳細な調査の実施',
-                '設定値の見直し',
-                'リソース監視の強化'
-            ]
-        };
-    }
-}
-
-// 基本的な検出器の実装（一部のみ）
-class FrameRateBottleneckDetector {
-    async detect(data, thresholds) {
-        const bottlenecks = [];
-        const frameRateThreshold = thresholds.get('frame_rate');
         
-        for (const sample of data.rawData.samples) {
-            if (sample.frameRate < frameRateThreshold.critical) {
-                bottlenecks.push({
-                    type: 'frame_rate',
-                    severity: 'critical',
-                    component: 'rendering_pipeline',
-                    description: `Critical frame rate drop: ${sample.frameRate.toFixed(1)}fps`,
-                    timestamp: sample.timestamp,
-                    value: sample.frameRate,
-                    threshold: frameRateThreshold.critical,
-                    impact: 'severe_user_experience_degradation'
-                });
-            } else if (sample.frameRate < frameRateThreshold.high) {
-                bottlenecks.push({
-                    type: 'frame_rate',
-                    severity: 'high',
-                    component: 'rendering_pipeline',
-                    description: `Frame rate below target: ${sample.frameRate.toFixed(1)}fps`,
-                    timestamp: sample.timestamp,
-                    value: sample.frameRate,
-                    threshold: frameRateThreshold.high,
-                    impact: 'noticeable_performance_impact'
-                });
-            }
+        if (config.reporting) {
+            this.reporter.configure(config.reporting);
         }
-
-        return this.deduplicateBottlenecks(bottlenecks);
-    }
-
-    deduplicateBottlenecks(bottlenecks) {
-        // 連続する同様のボトルネックを統合
-        const grouped = new Map();
         
-        for (const bottleneck of bottlenecks) {
-            const key = `${bottleneck.type}_${bottleneck.severity}`;
-            if (!grouped.has(key)) {
-                grouped.set(key, {
-                    ...bottleneck,
-                    occurrences: 1,
-                    firstSeen: bottleneck.timestamp,
-                    lastSeen: bottleneck.timestamp,
-                    minValue: bottleneck.value,
-                    maxValue: bottleneck.value
-                });
-            } else {
-                const existing = grouped.get(key);
-                existing.occurrences++;
-                existing.lastSeen = bottleneck.timestamp;
-                existing.minValue = Math.min(existing.minValue, bottleneck.value);
-                existing.maxValue = Math.max(existing.maxValue, bottleneck.value);
+        console.log('[PerformanceDiagnostics] Configuration updated');
+    }
+
+    /**
+     * Get component references for advanced usage
+     * @returns {Object} Component references
+     */
+    getComponents() {
+        return {
+            dataCollector: this.dataCollector,
+            analyzer: this.analyzer,
+            reporter: this.reporter
+        };
+    }
+
+    /**
+     * Cleanup diagnostic system
+     */
+    destroy() {
+        try {
+            // Destroy sub-components
+            if (this.dataCollector.destroy) {
+                this.dataCollector.destroy();
             }
+            
+            if (this.analyzer.destroy) {
+                this.analyzer.destroy();
+            }
+            
+            if (this.reporter.destroy) {
+                this.reporter.destroy();
+            }
+            
+            this.initialized = false;
+            console.log('[PerformanceDiagnostics] Main controller destroyed');
+        } catch (error) {
+            console.error('[PerformanceDiagnostics] Error during cleanup:', error);
         }
-
-        return Array.from(grouped.values());
     }
 }
 
-// 基本的な分析器の実装（スタブ）
-class MemoryBottleneckDetector {
-    async detect(data, thresholds) { return []; }
-}
+// Singleton instance
+let performanceDiagnosticsInstance = null;
 
-class RenderingBottleneckDetector {
-    async detect(data, thresholds) { return []; }
-}
-
-class NetworkBottleneckDetector {
-    async detect(data, thresholds) { return []; }
-}
-
-class ComputationBottleneckDetector {
-    async detect(data, thresholds) { return []; }
-}
-
-// 基本的な処理器の実装（スタブ）
-class FrameAnalysisProcessor {
-    async initialize() {}
-    async process(data) { return data; }
-}
-
-class MemoryAnalysisProcessor {
-    async initialize() {}
-    async process(data) { return data; }
-}
-
-class RenderAnalysisProcessor {
-    async initialize() {}
-    async process(data) { return data; }
-}
-
-class NetworkAnalysisProcessor {
-    async initialize() {}
-    async process(data) { return data; }
-}
-
-class InteractionAnalysisProcessor {
-    async initialize() {}
-    async process(data) { return data; }
-}
-
-// その他の基本クラス（スタブ）
-class DiagnosticRuleEngine {
-    async initialize() {}
-    async apply(data, context) { return data; }
-}
-
-class PerformanceDegradationAnalyzer {
-    async analyze(data, context) { return []; }
-}
-
-class ResourceContentionAnalyzer {
-    async analyze(data, context) { return []; }
-}
-
-class InefficientAlgorithmAnalyzer {
-    async analyze(data, context) { return []; }
-}
-
-class MemoryLeakAnalyzer {
-    async analyze(data, context) { return []; }
-}
-
-class ExcessiveGCAnalyzer {
-    async analyze(data, context) { return []; }
-}
-
-class OptimizationRecommendationGenerator {}
-class ConfigurationRecommendationGenerator {}
-class ArchitectureRecommendationGenerator {}
-class MonitoringRecommendationGenerator {}
-
-class PerformanceKnowledgeBase {
-    async initialize() {}
-}
-
-class ComprehensiveReportTemplate {
-    async generate(session) {
-        return {
-            type: 'comprehensive',
-            title: 'Comprehensive Performance Diagnostic Report',
-            session: session,
-            generatedAt: new Date().toISOString()
-        };
+/**
+ * Get PerformanceDiagnostics singleton instance
+ * @returns {PerformanceDiagnostics} Diagnostics instance
+ */
+export function getPerformanceDiagnostics() {
+    if (!performanceDiagnosticsInstance) {
+        performanceDiagnosticsInstance = new PerformanceDiagnostics();
     }
+    return performanceDiagnosticsInstance;
 }
 
-class SummaryReportTemplate {
-    async generate(session) {
-        return {
-            type: 'summary',
-            title: 'Performance Diagnostic Summary',
-            session: session,
-            generatedAt: new Date().toISOString()
-        };
-    }
-}
-
-class TechnicalReportTemplate {
-    async generate(session) {
-        return {
-            type: 'technical',
-            title: 'Technical Performance Analysis Report',
-            session: session,
-            generatedAt: new Date().toISOString()
-        };
-    }
-}
-
-class StatisticalAnomalyDetector {
-    async detect(data, baselines) { return []; }
-}
-
-class ThresholdAnomalyDetector {
-    async detect(data, baselines) { return []; }
-}
-
-class PatternAnomalyDetector {
-    async detect(data, baselines) { return []; }
-}
-
-class TrendAnomalyDetector {
-    async detect(data, baselines) { return []; }
-}
-
-class CorrelationAnalyzer {
-    async initialize() {}
-    async analyze(data) { return []; }
-}
-
-class CausalityDetector {
-    async initialize() {}
-    async detect(data, results) { return []; }
-}
-
-class DependencyMapper {
-    async initialize() {}
-    async map(results) { return []; }
-}
-
-export { PerformanceDiagnostics };
+export default PerformanceDiagnostics;
