+++ src/utils/performance-diagnostics/DiagnosticAnalyzer.js
@@ -1,0 +1,649 @@
+/**
+ * DiagnosticAnalyzer - Diagnostic analysis and problem detection functionality
+ * Part of the PerformanceDiagnostics split implementation
+ */
+
+export class DiagnosticAnalyzer {
+    constructor(mainController) {
+        this.mainController = mainController;
+        
+        // Analysis components
+        this.diagnosticEngine = new DiagnosticEngine();
+        this.bottleneckIdentifier = new BottleneckIdentifier();
+        this.issueAnalyzer = new IssueAnalyzer();
+        this.anomalyDetector = new AnomalyDetector();
+        this.rootCauseAnalyzer = new RootCauseAnalyzer();
+        
+        console.log('[DiagnosticAnalyzer] Analyzer component initialized');
+    }
+    
+    /**
+     * Initialize all analysis components
+     */
+    async initialize() {
+        try {
+            await this.diagnosticEngine.initialize();
+            await this.bottleneckIdentifier.initialize();
+            await this.issueAnalyzer.initialize();
+            await this.anomalyDetector.initialize();
+            await this.rootCauseAnalyzer.initialize();
+            
+            console.log('[DiagnosticAnalyzer] All analysis components initialized');
+        } catch (error) {
+            console.error('[DiagnosticAnalyzer] Failed to initialize analysis components:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Run comprehensive analysis on collected data
+     * @param {Object} collectedData - Collected diagnostic data
+     * @param {Object} options - Analysis options
+     * @returns {Object} Analysis results
+     */
+    async runAnalyses(collectedData, options) {
+        const results = {};
+        
+        // ボトルネック分析
+        if (options.includeBottleneckAnalysis) {
+            console.log('[DiagnosticAnalyzer] Running bottleneck analysis...');
+            results.bottlenecks = await this.bottleneckIdentifier.analyze(collectedData);
+        }
+        
+        // 異常検出
+        if (options.includeAnomalyDetection) {
+            console.log('[DiagnosticAnalyzer] Running anomaly detection...');
+            results.anomalies = await this.anomalyDetector.detect(collectedData);
+        }
+        
+        // 根本原因分析
+        if (options.includeRootCauseAnalysis) {
+            console.log('[DiagnosticAnalyzer] Running root cause analysis...');
+            results.rootCauses = await this.rootCauseAnalyzer.analyze(collectedData, results);
+        }
+        
+        return results;
+    }
+
+    /**
+     * Generate overall assessment based on analysis results
+     * @param {Object} analysisResults - Analysis results
+     * @returns {Object} Overall assessment
+     */
+    async generateOverallAssessment(analysisResults) {
+        const assessment = {
+            healthScore: 100,
+            performanceLevel: 'excellent',
+            criticalIssues: [],
+            majorConcerns: [],
+            improvements: [],
+            summary: ''
+        };
+
+        // ボトルネック評価
+        if (analysisResults.bottlenecks) {
+            const criticalBottlenecks = analysisResults.bottlenecks.filter(b => b.severity === 'critical');
+            const majorBottlenecks = analysisResults.bottlenecks.filter(b => b.severity === 'high');
+            
+            assessment.healthScore -= criticalBottlenecks.length * 20;
+            assessment.healthScore -= majorBottlenecks.length * 10;
+            
+            assessment.criticalIssues.push(...criticalBottlenecks.map(b => ({
+                type: 'bottleneck',
+                description: b.description,
+                component: b.component,
+                impact: b.impact
+            })));
+        }
+
+        // 異常評価
+        if (analysisResults.anomalies) {
+            const criticalAnomalies = analysisResults.anomalies.filter(a => a.severity === 'critical');
+            const majorAnomalies = analysisResults.anomalies.filter(a => a.severity === 'high');
+            
+            assessment.healthScore -= criticalAnomalies.length * 15;
+            assessment.healthScore -= majorAnomalies.length * 8;
+            
+            assessment.criticalIssues.push(...criticalAnomalies.map(a => ({
+                type: 'anomaly',
+                description: a.description,
+                metric: a.metric,
+                deviation: a.deviation
+            })));
+        }
+
+        // パフォーマンスレベル決定
+        assessment.healthScore = Math.max(0, assessment.healthScore);
+        if (assessment.healthScore >= 90) assessment.performanceLevel = 'excellent';
+        else if (assessment.healthScore >= 75) assessment.performanceLevel = 'good';
+        else if (assessment.healthScore >= 60) assessment.performanceLevel = 'fair';
+        else if (assessment.healthScore >= 40) assessment.performanceLevel = 'poor';
+        else assessment.performanceLevel = 'critical';
+
+        // サマリー生成
+        assessment.summary = this.generateAssessmentSummary(assessment);
+
+        return assessment;
+    }
+
+    /**
+     * Generate assessment summary text
+     * @param {Object} assessment - Assessment data
+     * @returns {string} Assessment summary
+     */
+    generateAssessmentSummary(assessment) {
+        const level = assessment.performanceLevel;
+        const score = assessment.healthScore;
+        const issues = assessment.criticalIssues.length;
+        
+        if (level === 'excellent') {
+            return `システムは優秀な状態です（スコア: ${score}）。大きな問題は検出されていません。`;
+        } else if (level === 'good') {
+            return `システムは良好な状態です（スコア: ${score}）。軽微な改善の余地があります。`;
+        } else if (level === 'fair') {
+            return `システムは普通の状態です（スコア: ${score}）。${issues}件の問題が検出されました。`;
+        } else if (level === 'poor') {
+            return `システムに性能問題があります（スコア: ${score}）。${issues}件の重要な問題への対処が必要です。`;
+        } else {
+            return `システムは深刻な状態です（スコア: ${score}）。${issues}件の致命的な問題への緊急対処が必要です。`;
+        }
+    }
+
+    /**
+     * Quick diagnosis for basic analysis
+     * @param {Object} collectedData - Collected data
+     * @returns {Object} Quick diagnosis results
+     */
+    async performQuickDiagnosis(collectedData) {
+        const options = {
+            includeBottleneckAnalysis: true,
+            includeAnomalyDetection: true,
+            includeRootCauseAnalysis: false
+        };
+
+        return await this.runAnalyses(collectedData, options);
+    }
+
+    /**
+     * Identify bottlenecks only
+     * @param {Object} collectedData - Collected data
+     * @returns {Array} Identified bottlenecks
+     */
+    async identifyBottlenecks(collectedData) {
+        return await this.bottleneckIdentifier.analyze(collectedData);
+    }
+
+    /**
+     * Detect anomalies only
+     * @param {Object} collectedData - Collected data
+     * @returns {Array} Detected anomalies
+     */
+    async detectAnomalies(collectedData) {
+        return await this.anomalyDetector.detect(collectedData);
+    }
+
+    /**
+     * Get analysis capabilities
+     * @returns {Object} Available analysis capabilities
+     */
+    getAnalysisCapabilities() {
+        return {
+            bottleneckAnalysis: true,
+            anomalyDetection: true,
+            rootCauseAnalysis: true,
+            performanceAssessment: true,
+            supportedMetrics: [
+                'frameRate',
+                'memoryUsage', 
+                'renderTime',
+                'networkLatency',
+                'inputLag'
+            ]
+        };
+    }
+
+    /**
+     * Configure analysis components
+     * @param {Object} config - Configuration options
+     */
+    configure(config) {
+        if (config.bottleneckThresholds) {
+            this.bottleneckIdentifier.updateThresholds(config.bottleneckThresholds);
+        }
+        
+        if (config.anomalyBaselines) {
+            this.anomalyDetector.updateBaselines(config.anomalyBaselines);
+        }
+        
+        console.log('[DiagnosticAnalyzer] Configuration updated');
+    }
+
+    /**
+     * Cleanup analyzer resources
+     */
+    destroy() {
+        // Cleanup analysis components if they have destroy methods
+        const components = [
+            this.diagnosticEngine,
+            this.bottleneckIdentifier,
+            this.issueAnalyzer,
+            this.anomalyDetector,
+            this.rootCauseAnalyzer
+        ];
+
+        components.forEach(component => {
+            if (component.destroy && typeof component.destroy === 'function') {
+                component.destroy();
+            }
+        });
+
+        console.log('[DiagnosticAnalyzer] Analyzer destroyed');
+    }
+}
+
+// 診断エンジン
+class DiagnosticEngine {
+    constructor() {
+        this.processors = new Map();
+        this.ruleEngine = new DiagnosticRuleEngine();
+    }
+
+    async initialize() {
+        await this.setupProcessors();
+        await this.ruleEngine.initialize();
+    }
+
+    async setupProcessors() {
+        this.processors.set('frame_analysis', new FrameAnalysisProcessor());
+        this.processors.set('memory_analysis', new MemoryAnalysisProcessor());
+        this.processors.set('render_analysis', new RenderAnalysisProcessor());
+        this.processors.set('network_analysis', new NetworkAnalysisProcessor());
+        this.processors.set('interaction_analysis', new InteractionAnalysisProcessor());
+
+        for (const processor of this.processors.values()) {
+            await processor.initialize();
+        }
+    }
+
+    async processData(data, analysisType) {
+        const processor = this.processors.get(analysisType);
+        if (!processor) {
+            throw new Error(`Unknown analysis type: ${analysisType}`);
+        }
+
+        return await processor.process(data);
+    }
+
+    async applyRules(data, context) {
+        return await this.ruleEngine.apply(data, context);
+    }
+}
+
+// ボトルネック特定器
+class BottleneckIdentifier {
+    constructor() {
+        this.detectors = new Map();
+        this.thresholds = new Map();
+    }
+
+    async initialize() {
+        this.setupDetectors();
+        this.setupThresholds();
+    }
+
+    setupDetectors() {
+        this.detectors.set('frame_rate', new FrameRateBottleneckDetector());
+        this.detectors.set('memory', new MemoryBottleneckDetector());
+        this.detectors.set('rendering', new RenderingBottleneckDetector());
+        this.detectors.set('network', new NetworkBottleneckDetector());
+        this.detectors.set('computation', new ComputationBottleneckDetector());
+    }
+
+    setupThresholds() {
+        this.thresholds.set('frame_rate', {
+            critical: 20, // 20fps以下は致命的
+            high: 40,     // 40fps以下は高優先度
+            medium: 55    // 55fps以下は中優先度
+        });
+
+        this.thresholds.set('memory_growth', {
+            critical: 5 * 1024 * 1024, // 5MB/s以上の成長は致命的
+            high: 2 * 1024 * 1024,     // 2MB/s以上は高優先度
+            medium: 1 * 1024 * 1024    // 1MB/s以上は中優先度
+        });
+
+        this.thresholds.set('render_time', {
+            critical: 50, // 50ms以上は致命的
+            high: 33.33,  // 30fps相当以上は高優先度
+            medium: 20    // 20ms以上は中優先度
+        });
+    }
+
+    updateThresholds(newThresholds) {
+        for (const [key, value] of Object.entries(newThresholds)) {
+            this.thresholds.set(key, value);
+        }
+    }
+
+    async analyze(collectedData) {
+        const bottlenecks = [];
+
+        for (const [detectorName, detector] of this.detectors) {
+            try {
+                const detected = await detector.detect(collectedData, this.thresholds);
+                bottlenecks.push(...detected);
+            } catch (error) {
+                console.error(`[BottleneckIdentifier] Detection failed for ${detectorName}:`, error);
+            }
+        }
+
+        // 重要度でソート
+        return bottlenecks.sort((a, b) => {
+            const severityOrder = { critical: 4, high: 3, medium: 2, low: 1 };
+            return (severityOrder[b.severity] || 0) - (severityOrder[a.severity] || 0);
+        });
+    }
+}
+
+// 問題分析器
+class IssueAnalyzer {
+    constructor() {
+        this.analyzers = new Map();
+    }
+
+    async initialize() {
+        this.analyzers.set('performance_degradation', new PerformanceDegradationAnalyzer());
+        this.analyzers.set('resource_contention', new ResourceContentionAnalyzer());
+        this.analyzers.set('inefficient_algorithms', new InefficientAlgorithmAnalyzer());
+        this.analyzers.set('memory_leaks', new MemoryLeakAnalyzer());
+        this.analyzers.set('excessive_garbage_collection', new ExcessiveGCAnalyzer());
+    }
+
+    async analyzeIssues(data, context) {
+        const issues = [];
+
+        for (const [analyzerName, analyzer] of this.analyzers) {
+            try {
+                const detected = await analyzer.analyze(data, context);
+                issues.push(...detected);
+            } catch (error) {
+                console.error(`[IssueAnalyzer] Analysis failed for ${analyzerName}:`, error);
+            }
+        }
+
+        return issues;
+    }
+}
+
+// 異常検出器
+class AnomalyDetector {
+    constructor() {
+        this.detectors = new Map();
+        this.baselines = new Map();
+    }
+
+    async initialize() {
+        this.setupDetectors();
+        this.loadBaselines();
+    }
+
+    setupDetectors() {
+        this.detectors.set('statistical', new StatisticalAnomalyDetector());
+        this.detectors.set('threshold', new ThresholdAnomalyDetector());
+        this.detectors.set('pattern', new PatternAnomalyDetector());
+        this.detectors.set('trend', new TrendAnomalyDetector());
+    }
+
+    loadBaselines() {
+        // ベースライン値の設定
+        this.baselines.set('frameRate', { mean: 60, stdDev: 5 });
+        this.baselines.set('memoryUsage', { mean: 50 * 1024 * 1024, stdDev: 10 * 1024 * 1024 });
+        this.baselines.set('renderTime', { mean: 12, stdDev: 3 });
+        this.baselines.set('networkLatency', { mean: 50, stdDev: 20 });
+        this.baselines.set('inputLag', { mean: 15, stdDev: 5 });
+    }
+
+    updateBaselines(newBaselines) {
+        for (const [key, value] of Object.entries(newBaselines)) {
+            this.baselines.set(key, value);
+        }
+    }
+
+    async detect(collectedData) {
+        const anomalies = [];
+
+        for (const [detectorName, detector] of this.detectors) {
+            try {
+                const detected = await detector.detect(collectedData, this.baselines);
+                anomalies.push(...detected);
+            } catch (error) {
+                console.error(`[AnomalyDetector] Detection failed for ${detectorName}:`, error);
+            }
+        }
+
+        return anomalies.sort((a, b) => {
+            const severityOrder = { critical: 4, high: 3, medium: 2, low: 1 };
+            return (severityOrder[b.severity] || 0) - (severityOrder[a.severity] || 0);
+        });
+    }
+}
+
+// 根本原因分析器
+class RootCauseAnalyzer {
+    constructor() {
+        this.correlationAnalyzer = new CorrelationAnalyzer();
+        this.causalityDetector = new CausalityDetector();
+        this.dependencyMapper = new DependencyMapper();
+    }
+
+    async initialize() {
+        await this.correlationAnalyzer.initialize();
+        await this.causalityDetector.initialize();
+        await this.dependencyMapper.initialize();
+    }
+
+    async analyze(collectedData, analysisResults) {
+        const rootCauses = [];
+
+        // 相関分析
+        const correlations = await this.correlationAnalyzer.analyze(collectedData);
+        
+        // 因果関係検出
+        const causalities = await this.causalityDetector.detect(collectedData, analysisResults);
+        
+        // 依存関係マッピング
+        const dependencies = await this.dependencyMapper.map(analysisResults);
+
+        // 根本原因の特定
+        for (const bottleneck of analysisResults.bottlenecks || []) {
+            const cause = await this.identifyRootCause(bottleneck, correlations, causalities, dependencies);
+            if (cause) {
+                rootCauses.push(cause);
+            }
+        }
+
+        return rootCauses;
+    }
+
+    async identifyRootCause(issue, correlations, causalities, dependencies) {
+        // 簡易的な根本原因特定
+        return {
+            issue: issue.description,
+            potentialCauses: [
+                'システム負荷の増大',
+                '不適切な設定値',
+                'リソースの枯渇',
+                'アルゴリズムの非効率性'
+            ],
+            confidence: 0.7,
+            recommendedActions: [
+                '詳細な調査の実施',
+                '設定値の見直し',
+                'リソース監視の強化'
+            ]
+        };
+    }
+}
+
+// 基本的な検出器の実装
+class FrameRateBottleneckDetector {
+    async detect(data, thresholds) {
+        const bottlenecks = [];
+        const frameRateThreshold = thresholds.get('frame_rate');
+        
+        for (const sample of data.rawData.samples) {
+            if (sample.frameRate < frameRateThreshold.critical) {
+                bottlenecks.push({
+                    type: 'frame_rate',
+                    severity: 'critical',
+                    component: 'rendering_pipeline',
+                    description: `Critical frame rate drop: ${sample.frameRate.toFixed(1)}fps`,
+                    timestamp: sample.timestamp,
+                    value: sample.frameRate,
+                    threshold: frameRateThreshold.critical,
+                    impact: 'severe_user_experience_degradation'
+                });
+            } else if (sample.frameRate < frameRateThreshold.high) {
+                bottlenecks.push({
+                    type: 'frame_rate',
+                    severity: 'high',
+                    component: 'rendering_pipeline',
+                    description: `Frame rate below target: ${sample.frameRate.toFixed(1)}fps`,
+                    timestamp: sample.timestamp,
+                    value: sample.frameRate,
+                    threshold: frameRateThreshold.high,
+                    impact: 'noticeable_performance_impact'
+                });
+            }
+        }
+
+        return this.deduplicateBottlenecks(bottlenecks);
+    }
+
+    deduplicateBottlenecks(bottlenecks) {
+        // 連続する同様のボトルネックを統合
+        const grouped = new Map();
+        
+        for (const bottleneck of bottlenecks) {
+            const key = `${bottleneck.type}_${bottleneck.severity}`;
+            if (!grouped.has(key)) {
+                grouped.set(key, {
+                    ...bottleneck,
+                    occurrences: 1,
+                    firstSeen: bottleneck.timestamp,
+                    lastSeen: bottleneck.timestamp,
+                    minValue: bottleneck.value,
+                    maxValue: bottleneck.value
+                });
+            } else {
+                const existing = grouped.get(key);
+                existing.occurrences++;
+                existing.lastSeen = bottleneck.timestamp;
+                existing.minValue = Math.min(existing.minValue, bottleneck.value);
+                existing.maxValue = Math.max(existing.maxValue, bottleneck.value);
+            }
+        }
+
+        return Array.from(grouped.values());
+    }
+}
+
+// 基本的な分析器の実装（スタブ）
+class MemoryBottleneckDetector {
+    async detect(data, thresholds) { return []; }
+}
+
+class RenderingBottleneckDetector {
+    async detect(data, thresholds) { return []; }
+}
+
+class NetworkBottleneckDetector {
+    async detect(data, thresholds) { return []; }
+}
+
+class ComputationBottleneckDetector {
+    async detect(data, thresholds) { return []; }
+}
+
+// 基本的な処理器の実装（スタブ）
+class FrameAnalysisProcessor {
+    async initialize() {}
+    async process(data) { return data; }
+}
+
+class MemoryAnalysisProcessor {
+    async initialize() {}
+    async process(data) { return data; }
+}
+
+class RenderAnalysisProcessor {
+    async initialize() {}
+    async process(data) { return data; }
+}
+
+class NetworkAnalysisProcessor {
+    async initialize() {}
+    async process(data) { return data; }
+}
+
+class InteractionAnalysisProcessor {
+    async initialize() {}
+    async process(data) { return data; }
+}
+
+// その他の基本クラス（スタブ）
+class DiagnosticRuleEngine {
+    async initialize() {}
+    async apply(data, context) { return data; }
+}
+
+class PerformanceDegradationAnalyzer {
+    async analyze(data, context) { return []; }
+}
+
+class ResourceContentionAnalyzer {
+    async analyze(data, context) { return []; }
+}
+
+class InefficientAlgorithmAnalyzer {
+    async analyze(data, context) { return []; }
+}
+
+class MemoryLeakAnalyzer {
+    async analyze(data, context) { return []; }
+}
+
+class ExcessiveGCAnalyzer {
+    async analyze(data, context) { return []; }
+}
+
+class StatisticalAnomalyDetector {
+    async detect(data, baselines) { return []; }
+}
+
+class ThresholdAnomalyDetector {
+    async detect(data, baselines) { return []; }
+}
+
+class PatternAnomalyDetector {
+    async detect(data, baselines) { return []; }
+}
+
+class TrendAnomalyDetector {
+    async detect(data, baselines) { return []; }
+}
+
+class CorrelationAnalyzer {
+    async initialize() {}
+    async analyze(data) { return []; }
+}
+
+class CausalityDetector {
+    async initialize() {}
+    async detect(data, results) { return []; }
+}
+
+class DependencyMapper {
+    async initialize() {}
+    async map(results) { return []; }
+}
