+++ src/utils/performance-diagnostics/DiagnosticReporter.js
@@ -1,0 +1,505 @@
+/**
+ * DiagnosticReporter - Diagnostic reporting and result presentation functionality
+ * Part of the PerformanceDiagnostics split implementation
+ */
+
+export class DiagnosticReporter {
+    constructor(mainController) {
+        this.mainController = mainController;
+        
+        // Reporting components
+        this.reportGenerator = new DiagnosticReportGenerator();
+        this.recommendationEngine = new RecommendationEngine();
+        
+        console.log('[DiagnosticReporter] Reporter component initialized');
+    }
+    
+    /**
+     * Initialize reporting components
+     */
+    async initialize() {
+        try {
+            await this.reportGenerator.initialize();
+            await this.recommendationEngine.initialize();
+            
+            console.log('[DiagnosticReporter] All reporting components initialized');
+        } catch (error) {
+            console.error('[DiagnosticReporter] Failed to initialize reporting components:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Generate comprehensive diagnostic report
+     * @param {Object} diagnosticSession - Diagnostic session data
+     * @returns {Object} Generated report
+     */
+    async generateReport(diagnosticSession) {
+        return await this.reportGenerator.generate(diagnosticSession);
+    }
+    
+    /**
+     * Generate recommendations based on analysis results
+     * @param {Object} analysisResults - Analysis results
+     * @returns {Array} Generated recommendations
+     */
+    async generateRecommendations(analysisResults) {
+        return await this.recommendationEngine.generate(analysisResults);
+    }
+    
+    /**
+     * Generate quick summary report
+     * @param {Object} diagnosticSession - Diagnostic session data
+     * @returns {Object} Summary report
+     */
+    async generateSummaryReport(diagnosticSession) {
+        const summary = {
+            sessionId: diagnosticSession.id,
+            duration: diagnosticSession.duration,
+            timestamp: new Date().toISOString(),
+            overallHealth: diagnosticSession.results.overallAssessment?.healthScore || 0,
+            performanceLevel: diagnosticSession.results.overallAssessment?.performanceLevel || 'unknown',
+            criticalIssues: diagnosticSession.results.overallAssessment?.criticalIssues?.length || 0,
+            bottlenecks: diagnosticSession.results.bottlenecks?.length || 0,
+            anomalies: diagnosticSession.results.anomalies?.length || 0,
+            recommendations: diagnosticSession.results.recommendations?.length || 0
+        };
+        
+        return {
+            type: 'summary',
+            title: 'Performance Diagnostic Summary',
+            summary: summary,
+            generatedAt: new Date().toISOString()
+        };
+    }
+    
+    /**
+     * Generate detailed technical report
+     * @param {Object} diagnosticSession - Diagnostic session data
+     * @returns {Object} Technical report
+     */
+    async generateTechnicalReport(diagnosticSession) {
+        const technicalDetails = {
+            dataCollection: {
+                duration: diagnosticSession.results.dataCollection?.duration || 0,
+                sampleCount: diagnosticSession.results.dataCollection?.sampleCount || 0,
+                metricsCollected: diagnosticSession.results.dataCollection?.metricsCollected || 0,
+                dataQuality: diagnosticSession.results.dataCollection?.dataQuality || {}
+            },
+            bottleneckAnalysis: {
+                totalBottlenecks: diagnosticSession.results.bottlenecks?.length || 0,
+                criticalBottlenecks: diagnosticSession.results.bottlenecks?.filter(b => b.severity === 'critical').length || 0,
+                categories: this.categorizeIssues(diagnosticSession.results.bottlenecks || [])
+            },
+            anomalyDetection: {
+                totalAnomalies: diagnosticSession.results.anomalies?.length || 0,
+                criticalAnomalies: diagnosticSession.results.anomalies?.filter(a => a.severity === 'critical').length || 0,
+                detectionTypes: this.categorizeAnomalies(diagnosticSession.results.anomalies || [])
+            },
+            systemAssessment: diagnosticSession.results.overallAssessment || {}
+        };
+        
+        return {
+            type: 'technical',
+            title: 'Technical Performance Analysis Report',
+            technicalDetails: technicalDetails,
+            rawResults: diagnosticSession.results,
+            generatedAt: new Date().toISOString()
+        };
+    }
+    
+    /**
+     * Categorize issues by type
+     * @param {Array} issues - Array of issues
+     * @returns {Object} Categorized issues
+     */
+    categorizeIssues(issues) {
+        const categories = {};
+        
+        issues.forEach(issue => {
+            const category = issue.type || 'unknown';
+            categories[category] = (categories[category] || 0) + 1;
+        });
+        
+        return categories;
+    }
+    
+    /**
+     * Categorize anomalies by detection type
+     * @param {Array} anomalies - Array of anomalies
+     * @returns {Object} Categorized anomalies
+     */
+    categorizeAnomalies(anomalies) {
+        const categories = {};
+        
+        anomalies.forEach(anomaly => {
+            const category = anomaly.detectionType || 'unknown';
+            categories[category] = (categories[category] || 0) + 1;
+        });
+        
+        return categories;
+    }
+    
+    /**
+     * Format report for display
+     * @param {Object} report - Report data
+     * @param {string} format - Output format ('text', 'html', 'json')
+     * @returns {string} Formatted report
+     */
+    formatReport(report, format = 'text') {
+        switch (format) {
+            case 'html':
+                return this.formatReportAsHTML(report);
+            case 'json':
+                return JSON.stringify(report, null, 2);
+            case 'text':
+            default:
+                return this.formatReportAsText(report);
+        }
+    }
+    
+    /**
+     * Format report as text
+     * @param {Object} report - Report data
+     * @returns {string} Text formatted report
+     */
+    formatReportAsText(report) {
+        const lines = [];
+        
+        lines.push(`=== ${report.title || 'Diagnostic Report'} ===`);
+        lines.push(`Generated: ${report.generatedAt}`);
+        lines.push('');
+        
+        if (report.summary) {
+            lines.push('Summary:');
+            lines.push(`  Health Score: ${report.summary.overallHealth}/100`);
+            lines.push(`  Performance Level: ${report.summary.performanceLevel}`);
+            lines.push(`  Critical Issues: ${report.summary.criticalIssues}`);
+            lines.push(`  Total Bottlenecks: ${report.summary.bottlenecks}`);
+            lines.push(`  Total Anomalies: ${report.summary.anomalies}`);
+            lines.push('');
+        }
+        
+        if (report.technicalDetails) {
+            lines.push('Technical Details:');
+            lines.push(`  Data Collection Duration: ${report.technicalDetails.dataCollection.duration}ms`);
+            lines.push(`  Samples Collected: ${report.technicalDetails.dataCollection.sampleCount}`);
+            lines.push(`  Metrics Collected: ${report.technicalDetails.dataCollection.metricsCollected}`);
+            lines.push('');
+        }
+        
+        return lines.join('\n');
+    }
+    
+    /**
+     * Format report as HTML
+     * @param {Object} report - Report data
+     * @returns {string} HTML formatted report
+     */
+    formatReportAsHTML(report) {
+        let html = `<div class="diagnostic-report">`;
+        html += `<h1>${report.title || 'Diagnostic Report'}</h1>`;
+        html += `<p class="generated-at">Generated: ${report.generatedAt}</p>`;
+        
+        if (report.summary) {
+            html += `<div class="summary">`;
+            html += `<h2>Summary</h2>`;
+            html += `<ul>`;
+            html += `<li>Health Score: ${report.summary.overallHealth}/100</li>`;
+            html += `<li>Performance Level: ${report.summary.performanceLevel}</li>`;
+            html += `<li>Critical Issues: ${report.summary.criticalIssues}</li>`;
+            html += `<li>Total Bottlenecks: ${report.summary.bottlenecks}</li>`;
+            html += `<li>Total Anomalies: ${report.summary.anomalies}</li>`;
+            html += `</ul>`;
+            html += `</div>`;
+        }
+        
+        html += `</div>`;
+        return html;
+    }
+    
+    /**
+     * Get reporting capabilities
+     * @returns {Object} Available reporting capabilities
+     */
+    getReportingCapabilities() {
+        return {
+            reportTypes: ['comprehensive', 'summary', 'technical'],
+            outputFormats: ['text', 'html', 'json'],
+            recommendationTypes: ['optimization', 'configuration', 'architecture', 'monitoring'],
+            customTemplates: true,
+            realTimeReporting: true
+        };
+    }
+    
+    /**
+     * Configure reporting components
+     * @param {Object} config - Configuration options
+     */
+    configure(config) {
+        if (config.reportTemplate) {
+            this.reportGenerator.setTemplate(config.reportTemplate);
+        }
+        
+        if (config.recommendationSettings) {
+            this.recommendationEngine.configure(config.recommendationSettings);
+        }
+        
+        console.log('[DiagnosticReporter] Configuration updated');
+    }
+    
+    /**
+     * Cleanup reporter resources
+     */
+    destroy() {
+        if (this.reportGenerator.destroy) {
+            this.reportGenerator.destroy();
+        }
+        
+        if (this.recommendationEngine.destroy) {
+            this.recommendationEngine.destroy();
+        }
+        
+        console.log('[DiagnosticReporter] Reporter destroyed');
+    }
+}
+
+// 推奨事項エンジン
+class RecommendationEngine {
+    constructor() {
+        this.generators = new Map();
+        this.knowledgeBase = new PerformanceKnowledgeBase();
+    }
+
+    async initialize() {
+        await this.knowledgeBase.initialize();
+        this.setupGenerators();
+    }
+
+    setupGenerators() {
+        this.generators.set('optimization', new OptimizationRecommendationGenerator());
+        this.generators.set('configuration', new ConfigurationRecommendationGenerator());
+        this.generators.set('architecture', new ArchitectureRecommendationGenerator());
+        this.generators.set('monitoring', new MonitoringRecommendationGenerator());
+    }
+
+    async generate(analysisResults) {
+        const recommendations = [];
+
+        // ボトルネックベースの推奨事項
+        if (analysisResults.bottlenecks) {
+            for (const bottleneck of analysisResults.bottlenecks) {
+                const rec = await this.generateBottleneckRecommendations(bottleneck);
+                recommendations.push(...rec);
+            }
+        }
+
+        // 異常ベースの推奨事項
+        if (analysisResults.anomalies) {
+            for (const anomaly of analysisResults.anomalies) {
+                const rec = await this.generateAnomalyRecommendations(anomaly);
+                recommendations.push(...rec);
+            }
+        }
+
+        // 一般的な推奨事項
+        const generalRec = await this.generateGeneralRecommendations(analysisResults);
+        recommendations.push(...generalRec);
+
+        return this.prioritizeAndDeduplicateRecommendations(recommendations);
+    }
+
+    async generateBottleneckRecommendations(bottleneck) {
+        const recommendations = [];
+
+        switch (bottleneck.type) {
+            case 'frame_rate':
+                recommendations.push({
+                    type: 'optimization',
+                    priority: 'high',
+                    category: 'rendering',
+                    title: 'フレームレート最適化',
+                    description: 'レンダリングパイプラインの最適化によりフレームレートを改善',
+                    actions: [
+                        'AdaptiveQualityController の設定確認',
+                        'レンダリング負荷の軽減',
+                        'フレームペーシングの調整',
+                        'VSync設定の最適化'
+                    ],
+                    estimatedImpact: 'high',
+                    implementationEffort: 'medium',
+                    timeToImplement: '1-3 days'
+                });
+                break;
+
+            case 'memory':
+                recommendations.push({
+                    type: 'memory',
+                    priority: 'high',
+                    category: 'memory_management',
+                    title: 'メモリ使用量最適化',
+                    description: 'メモリ管理の改善により安定性を向上',
+                    actions: [
+                        'MemoryManager の設定見直し',
+                        'オブジェクトプールの効率化',
+                        'メモリリークの修正',
+                        'ガベージコレクション最適化'
+                    ],
+                    estimatedImpact: 'high',
+                    implementationEffort: 'medium',
+                    timeToImplement: '2-5 days'
+                });
+                break;
+
+            case 'rendering':
+                recommendations.push({
+                    type: 'rendering',
+                    priority: 'medium',
+                    category: 'rendering_optimization',
+                    title: 'レンダリング効率改善',
+                    description: 'レンダリング処理の最適化',
+                    actions: [
+                        'ダーティリージョン管理の改善',
+                        'バッチレンダリングの実装',
+                        'レイヤー構成の最適化',
+                        'シェーダー最適化'
+                    ],
+                    estimatedImpact: 'medium',
+                    implementationEffort: 'medium',
+                    timeToImplement: '3-7 days'
+                });
+                break;
+        }
+
+        return recommendations;
+    }
+
+    async generateAnomalyRecommendations(anomaly) {
+        return [{
+            type: 'investigation',
+            priority: 'medium',
+            category: 'anomaly_resolution',
+            title: `${anomaly.metric}異常の調査`,
+            description: `${anomaly.metric}で検出された異常パターンの調査と対策`,
+            actions: [
+                '異常発生パターンの詳細分析',
+                '関連システムとの相関調査',
+                '閾値設定の見直し',
+                '予防的監視の強化'
+            ],
+            estimatedImpact: 'medium',
+            implementationEffort: 'low',
+            timeToImplement: '1-2 days'
+        }];
+    }
+
+    async generateGeneralRecommendations(analysisResults) {
+        return [
+            {
+                type: 'monitoring',
+                priority: 'low',
+                category: 'continuous_improvement',
+                title: '継続的パフォーマンス監視',
+                description: '定期的なパフォーマンス監視体制の確立',
+                actions: [
+                    'PerformanceMonitoringSystem の定期実行',
+                    'アラート閾値の定期見直し',
+                    'パフォーマンストレンドの分析',
+                    'ベースライン性能の更新'
+                ],
+                estimatedImpact: 'low',
+                implementationEffort: 'low',
+                timeToImplement: '1 day'
+            }
+        ];
+    }
+
+    prioritizeAndDeduplicateRecommendations(recommendations) {
+        // 重複除去
+        const unique = recommendations.filter((rec, index, array) => 
+            array.findIndex(r => r.title === rec.title) === index
+        );
+
+        // 優先度でソート
+        const priorityOrder = { high: 3, medium: 2, low: 1 };
+        return unique.sort((a, b) => 
+            (priorityOrder[b.priority] || 0) - (priorityOrder[a.priority] || 0)
+        );
+    }
+
+    configure(config) {
+        // Configure recommendation engine settings
+        console.log('[RecommendationEngine] Configuration updated');
+    }
+}
+
+// 診断レポート生成器
+class DiagnosticReportGenerator {
+    constructor() {
+        this.templates = new Map();
+    }
+
+    async initialize() {
+        this.setupTemplates();
+    }
+
+    setupTemplates() {
+        this.templates.set('comprehensive', new ComprehensiveReportTemplate());
+        this.templates.set('summary', new SummaryReportTemplate());
+        this.templates.set('technical', new TechnicalReportTemplate());
+    }
+
+    async generate(diagnosticSession) {
+        const templateName = diagnosticSession.options.detailLevel || 'comprehensive';
+        const template = this.templates.get(templateName) || this.templates.get('comprehensive');
+
+        return await template.generate(diagnosticSession);
+    }
+
+    setTemplate(templateName, template) {
+        this.templates.set(templateName, template);
+    }
+}
+
+// 基本クラス実装
+class OptimizationRecommendationGenerator {}
+class ConfigurationRecommendationGenerator {}
+class ArchitectureRecommendationGenerator {}
+class MonitoringRecommendationGenerator {}
+
+class PerformanceKnowledgeBase {
+    async initialize() {}
+}
+
+class ComprehensiveReportTemplate {
+    async generate(session) {
+        return {
+            type: 'comprehensive',
+            title: 'Comprehensive Performance Diagnostic Report',
+            session: session,
+            generatedAt: new Date().toISOString()
+        };
+    }
+}
+
+class SummaryReportTemplate {
+    async generate(session) {
+        return {
+            type: 'summary',
+            title: 'Performance Diagnostic Summary',
+            session: session,
+            generatedAt: new Date().toISOString()
+        };
+    }
+}
+
+class TechnicalReportTemplate {
+    async generate(session) {
+        return {
+            type: 'technical',
+            title: 'Technical Performance Analysis Report',
+            session: session,
+            generatedAt: new Date().toISOString()
+        };
+    }
+}
